requirements.txt
appdirs==1.4.4
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
idna==2.10
isodate==0.6.0
lxml==4.6.2
pycryptodome==3.9.9
pyDes==2.0.1
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.26.3
xmltodict==0.12.0
zeep==4.0.0
